TSS: Introduce TPM_NODEV for use in skiboot
Skiboot doesn't implement TPM device communication by opening and closing
the tpm device in /dev as linux does, instead it calls the tpmp driver
directly. So we introduce this CFLAG to rule out some unnecessary steps in
TPM communication workflow and the TSS_CONTEXT struct's dev_fd member, since we
don't use it for skiboot.

For instance we don't need to close the device in TSS_Close.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>
Signed-off-by: Mauro S. M. Rodrigues <maurosr@linux.vnet.ibm.com>