TSS/Eventlib: ifdef out non implemeted functions in skiboot
Skiboot doesn't implement fread, fwrite and sscanf and thus we need to
ifdef out functions or part of functions that usem these.

For TSS_EVENT2_Line_Read and TSS_EVENT_Line_Read both aren't used by
skiboot so it's easy to rule them out, but instead of relying in
TPM_SKIBOOT macro it makes a lot more sense to depend on the
TPM_TSS_NOFILE - which is set when compiling skiboot,

For TSS_SetCommandPort, TSS_SetPlatformPort,
TSS_SetEncryptSessions and TSS_Array_Scan it's also harmless to rule out
the piece of code using sscanf as it was done for __ULTRAVISOR__ in
commit c8bac25 ("TSS: ifdef out sscanf()").

And finally sscanf use in TSS_SetTraceLevel is ifdef out when
TPM_TSS_NO_PRINT is set.

Signed-off-by: Mauro S. M. Rodrigues <maurosr@linux.vnet.ibm.com>