TSS/Evenlib: ifdef out for TPM_SKIBOOT using non implemeted functions
Skiboot doesn't implement fread, fwrite or sscanf and thus we need to
ifdef out functions or part of functions that usem these.

For TSS_EVENT2_Line_Read and TSS_EVENT_Line_Read both aren't used by
skiboot so it's easy to rule them out.

For TSS_SetTraceLevel, TSS_SetCommandPort, TSS_SetPlatformPort,
TSS_SetEncryptSessions and TSS_Array_Scan it's also harmless to rule out
the piece of code using sscanf as it was done for __ULTRAVISOR__ in
commit c8bac25 ("TSS: ifdef out sscanf()").

Signed-off-by: Mauro S. M. Rodrigues <maurosr@linux.vnet.ibm.com>